+++ src/constants/AppConstants.js
@@ -1,3 +1,2 @@
 // Site
-export const FEEDBACK_URL = '/';
 export const SERVICE_NAME = 'National Maritime Single Window';
