+++ src/constants/AppUrlConstants.js
@@ -1,3 +1,7 @@
+// External links
+export const CROWN_COPYRIGHT_URL = 'https://www.nationalarchives.gov.uk/information-management/re-using-public-sector-information/uk-government-licensing-framework/crown-copyright/';
+export const FEEDBACK_URL = '/';
+
 // Pages
 export const COOKIE_URL = '/cookies';
 export const ACCESSIBILITY_URL = '/accessibility-statement';
